MiniEShop.Tests/Controllers/HomeControllerTest.cs
using System.Web.Mvc;
using BLL.BL;
using DomainClasses.Models;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using MiniEShop.Controllers;

namespace MiniEShop.Tests.Controllers
{
    [TestClass]
    public class HomeControllerTest
    {
        [TestMethod]
        public void Index()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.Index() as ViewResult;

            // Assert
            Assert.IsNotNull(result);

            ProductBL pBL = new ProductBL();


            Assert.IsTrue(pBL.Save(new Product() {ProductName = "Test"}));            

        }

        [TestMethod]
        public void About()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.About() as ViewResult;

            // Assert
            Assert.AreEqual("Your application description page.", result.ViewBag.Message);
        }

        [TestMethod]
        public void Contact()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.Contact() as ViewResult;

            // Assert
            Assert.IsNotNull(result);
        }
    }
}
